requirements.txt
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
certifi==2024.7.4
Django==5.0.6
django-debug-toolbar==4.4.3
djangorestframework==3.15.2
djangorestframework-oauth==1.1.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
PyJWT==2.8.0
python-dotenv==1.0.1
python-telegram-bot==21.3
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sniffio==1.3.1
sqlparse==0.5.0
uritemplate==4.1.1
